商品削除2

+++ app/controllers/items_controller.rb
@@ -1,6 +1,6 @@
 class ItemsController < ApplicationController
     before_action :authenticate_user!, except: [:index, :show]
-    before_action :set_item, only: [:show, :edit, :update, ]
+   # before_action :set_item, only: [:show, :edit, :update, ]
 
 def index
     @items = Item.all.order("created_at DESC")
@@ -28,29 +28,29 @@ def show
     #@orders = Order.find_by(params[:item_id])
 end
 
-def edit
- if user_signed_in? && current_user.id != @item.user_id 
-    redirect_to root_path
-    end
+#def edit
+ #if user_signed_in? && current_user.id != @item.user_id 
+  #  redirect_to root_path
+   # end
     
- unless user_signed_in?
-    redirect_to new_user_session_path
- end
-end
+ #unless user_signed_in?
+  #  redirect_to new_user_session_path
+ #end
+#end
 
-def update
-    @item.update(item_params)
-    if  @item.valid?
-        redirect_to user_item_path(@item.user_id)
-    else 
-        render :edit
+#def update
+ #   @item.update(item_params)
+  #  if  @item.valid?
+   #     redirect_to user_item_path(@item.user_id)
+    #else 
+     #   render :edit
     
-    end
-end
+   # end
+#end
 
 def destroy
     item = Item.find(params[:id])
-    if user_signed_in?
+    if if user_signed_in? && current_user.id != @item.user_id 
      item.destroy
     redirect_to action: :index
     end
